requirements.txt
asttokens==2.0.8
async-timeout==4.0.2
backcall==0.2.0
bidict==0.22.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.5
cycler==0.11.0
Deprecated==1.2.13
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
fonttools==4.38.0
idna==3.4
ipykernel==6.16.2
ipython
jedi==0.18.1
Jinja2==3.0.3
joblib==1.2.0
jsonpickle==2.2.0
jupyter-client==7.4.4
jupyter-core==4.11.2
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.0.1
matplotlib
matplotlib-inline
miditoolkit==0.1.16
mido==1.2.10
more-itertools==9.0.0
music21
muspy
nest-asyncio==1.5.6
numpy
packaging==21.3
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
pandas
Pillow
pretty-midi==0.2.9
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.8.2
Pygments==2.13.0
pyparsing==3.0.9
pypianoroll==1.0.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
redis==4.3.4
requests==2.28.1
scipy
six==1.16.0
spacy
spacy-legacy==3.0.8
spacy-loggers==1.0.1
spotipy
stack-data==0.5.1
thinc==8.0.13
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
webcolors==1.12
wincertstore==0.2
wrapt==1.14.1
